Enforce numerically plane stress condition for shell formulation (#125)
This pull request includes significant updates to the `initialization`
method in the `ShellStressRelaxationFirstHalf` class to improve the
handling of plane stress conditions. The changes involve replacing the
previous approach with a new iterative algorithm.

### Handling Plane Stress Conditions:

Replaced the `getCorrectedAlmansiStrain` method with a new iterative
algorithm to enforce plane stress conditions. This includes a hardcoded
tolerance and iteration limit, and throws an exception if the algorithm
fails to converge.

The algorithm follows the standard method for finite strain described in
Sec. 5.4.1 of the PhD thesis _"Differential geometry and the
geometrically non-linear Reissner-Mindlin shell model"_
[[doi]](http://dx.doi.org/10.18419/opus-14215), with 2 main adaptation.

1. The algorithm enforces $`\boldsymbol\sigma(\bf{e})=0`$ instead of
$`\bf{S}(\bf{E})=0`$ due to existing implementation choices of SPHinXsys
2. The algorithm uses the [secant
method](https://en.wikipedia.org/wiki/Secant_method), therefore relying
only on 1st order (i.e. strain and stress) information to avoid
implementing the calculation of the stiffness tangent. The tangent
update is calculated as

```math
E^{i+1}=\frac{\sigma^{i+1}-\sigma^{i}}{\bf{e}^{i+1}-\bf{e}^{i}}
```

where $E$ is the instantaneous Young's modulus.

The default algorithm, albeit a definitive improvement over the "linear
isotropic material" approximation, is not guaranteed to converge.
Indeed, the update procedure Eq. (5.127) propagates violation of the
positive definiteness, and thus, may fail for material model calculating
the stretches from the Left/Right Cauchy-Green tensor (wherein
$`\lambda_i^2\leq0`$).

Those additional modification obviously can render the algorithm more
brittle, checks are performed to detect non-finite values slipping in
the stress.

In past experiments with customer cases, the simulation was failing for
a set of parameters previously used, however the behavior of the
structure was visibly extreme (buckling, kinks) and not desired, pushing
for modifying simulation parameters rather than accepting it as an
intrinsic failure of the algorithm.

An improvement to the algorithm is described in the thesis mentioned
above in sec. 5.4.2, but its implementation is not straightforward due
to SPHinXsys peculiarities and left for a future issue.